requirements.txt
python-telegram-bot>=20.7,<22
requests>=2.31.0,<3
python-dotenv>=1.0.1,<2
